% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BVS4GCR.R
\name{BVS4GCR}
\alias{BVS4GCR}
\title{BVS4GCR}
\usage{
BVS4GCR(
  Y,
  X,
  pFixed,
  responseType,
  EVgamma = c(2, 4),
  tau = 1,
  niter,
  burnin,
  thin,
  monitor,
  fullCov = FALSE,
  nonDecomp = FALSE,
  nTrial = NULL,
  negBinomParInit = NULL,
  link = "probit",
  alpha = 0.05,
  probAddDel = 0.9,
  probAdd = 0.5,
  probMix = 0.25,
  geomMean = EVgamma[1],
  graphParNiter = max(16, dim(Y)[2] * dim(Y)[2]),
  std = FALSE,
  seed = 31122021
)
}
\arguments{
\item{Y}{(number of samples) times (number of responses) matrix of responses}

\item{X}{(number of samples) times (number of predictors) matrix of predictors. If the model contains an intercept (for all
responses), the first column is a vector of ones, followed by the covariates and the predictors on which Bayesian Variable
Selection is performed}

\item{pFixed}{Number of covariates (including the intercept)}

\item{responseType}{Character vector specifying the response's type. Response's types currently supported are \code{c("Gaussian", 
"binary", "ordinal", "binomial", "geometric", "negative binomial")}}

\item{EVgamma}{A priori number of expected predictors associated with each response and its variance. For details, see
\insertCite{Kohn2001;textual}{BVS4GCR} and \insertCite{Alexopoulos2021;textual}{BVS4GCR}. Default value is \code{EVgamma = c(2, 2)}
which implies a priori range of important predictors between 0 and 6 for each response}

\item{tau}{Prior variance of the beta (regression coefficient) prior with the default value set at \code{1}. If
\code{std = TRUE}, then \code{tau = 1}}

\item{niter}{Number of MCMC iterations (including burn-in)}

\item{burnin}{Number of iterations to be discarded as burn-in}

\item{thin}{Store the outcome at every thin-th iteration}

\item{monitor}{Display the monitored-th iteration}

\item{fullCov}{Logical parameter to select full or sparse inverse correlation}

\item{nonDecomp}{Logical, \code{TRUE} if a non-decomposable graphical model is selected}

\item{nTrial}{Number of trials of binomial responses if \code{"binomial"} is included in \code{responseType}}

\item{negBinomParInit}{Initial value for the over-dispersion parameter of the negative binomial distribution}

\item{link}{Specification for the model link function with default value (\code{"probit"})}

\item{alpha}{Level of significance (\code{0.05} default) to select important predictors for each response by using one at-a-time
response, univariable LM or GLM method. Used in the construction of the proposal distribution of \code{gamma}}

\item{probAddDel}{Probability (\code{0.9} default) of adding/deleting one predictor to be directly associated in a model with a
response during Markov chain Monte Carlo exploration}

\item{probAdd}{Probability (\code{0.9} default) of adding one predictor to be directly associated in a model with a response during
Markov chain Monte Monte Carlo exploration}

\item{probMix}{Probability (\code{0.25} default) of selecting a geometric proposal distribution that samples the index of the
predictor being added/deleted. For details, see \insertCite{Alexopoulos2021;textual}{BVS4GCR}}

\item{geomMean}{Mean of the geometric proposal distribution that samples the index of the predictor being added/deleted. Default
value is the number of expected predictors directly associated with each response and specified in \code{EVgamma}. For details,
see \insertCite{Alexopoulos2021;textual}{BVS4GCR}}

\item{graphParNiter}{Number of Markov chain Monte Carlo internal iterations to sample the graphical model between responses.
Default value is the max between 16 and the square of the number of responses}

\item{std}{Logical parameter to standardise the predictors before the analysis. Default value is set at \code{TRUE}}

\item{seed}{Seed used to initialise the Markov chain Monte Carlo algorithm}
}
\value{
The value returned is a list object \code{list(B, G, R, D, Cutoff, NBPar)}
\itemize{
\item{\code{Β}}{ Matrix of the (thinned) samples drawn from the posterior distribution of the regression coefficients}
\item{\code{G}}{ 3D array of the (thinned) samples drawn from the posterior distribution of the graphs}
\item{\code{R}}{ 3D array of the (thinned) samples drawn from the posterior distribution of the correlation matrix between the
responses}
\item{\code{D}}{ Matrix of the (thinned) samples drawn from the posterior distribution of the standard deviations of the
responses (non-identifiable parameters in the case of discrete responses)}
\item{\code{cutPoint}}{ List of the samples drawn from the posterior distribution of the cut-off points for ordinal categorical
responses}
\item{\code{NBPar}}{ List of the samples drawn from the posterior distribution of the over-dispersion parameter for the negative
responses}
\item{\code{postMean}}{ List of the posterior means of \code{list(B, G, R, D)} }
\item{\code{hyperPar}}{ List of the hyper-parameters \code{list(tau)} and the parameters of the beta-binomial distribution
derived from \code{EVgamma} }
\item{\code{samplerPar}}{ List of parameters used in the  Markov chain Monte Carlo algorithm
\code{list(alpha, probAddDel, probAdd, probMix, geomMean, graphParNiter)} }
\item{\code{opt}}{ List of options used \code{list(std, seed)} } }
}
\description{
MCMC implementation of Bayesian Variable Selection for Gaussian Copula Regression models
}
\details{
For details regarding the model and the algorithm, see details in \insertCite{Alexopoulos2021;textual}{BVS4GCR}. Types of
responses currently supported: \code{c("Gaussian", "binary", "ordinal", "binomial", "negative binomial")}. For ordinal
categorical responses it is assumed that the first category is labelled with zero. The maximum number of categories
supported is currently 5.
}
\examples{
# Example 1: Toy example with only Gaussian responses and decomposable graphs specification

d <- Sim_GCRdata(m = 4, n = 500, p = 100, pFixed = 2, responseType = rep("Gaussian", 4), 
                 extras = list(rhoY = 0.8, rhoX = 0.7, fixedEffect = rep(1, 2), muEffect = 
                               rep(3, 4), varEffect = rep(0.5, 4), s1Lev = 0.05, s2Lev = 0.95, 
                               sdResponse = rep(1, 4)), 
                 seed = 28061971)

output <- BVS4GCR(d$Y, d$X, pFixed = 2, responseType = rep("Gaussian", 4), EVgamma = c(5, 3^2), 
                  niter = 250, burnin = 50, thin = 4, monitor = 50, seed = 280610971)

# Example 2: Toy example with only Gaussian responses and non-decomposable graphs 

d <- Sim_GCRdata(m = 8, n = 1000, p = 100, pFixed = 1, responseType = rep("Gaussian", 8), 
                 extras = list(rhoY = 0.8, rhoX = 0.7, fixedEffect = 1, muEffect = rep(3, 8), 
                               varEffect = rep(0.5, 8), s1Lev = 0.10, s2Lev = 0.90, 
                               sdResponse = rep(1, 8)), 
                 seed = 28061971)

output <- BVS4GCR(d$Y, d$X, pFixed = 1, responseType = rep("Gaussian", 8), EVgamma = c(5, 3^2), 
                  niter = 250, burnin = 50, thin = 4, monitor = 50, 
                  nonDecomp = TRUE, seed = 280610971)


# Example 3: Combination of Gaussian, binary and ordinal responses (similar to Scenario I in
# Alexopoulos and Bottolo (2021))

d <- Sim_GCRdata(m = 4, n = 50, p = 30, pFixed = 1, responseType = c("Gaussian", "binary", 
                 "ordinal", "ordinal"), 
                 extras = list(rhoY = 0.8, rhoX = 0.7, fixedEffect = 1, muEffect = 
                               c(1, rep(0.5, 3)), varEffect = c(1, rep(0.2, 3)), s1Lev = 0.15, 
                               s2Lev = 0.95, sdResponse = rep(1, 4), nCat = c(3, 4), cutPoint = 
                               cbind(c(0.5, NA, NA), c(0.5, 1.2, 2))), 
                 seed = 28061971)

output <- BVS4GCR(d$Y, d$X, pFixed = 1, responseType = c("Gaussian", "binary", "ordinal", 
                  "ordinal"), EVgamma =  c(5, 3^2), 
                  niter = 250, burnin = 50, thin = 4, monitor = 50, seed = 280610971)


# Example 4: Combination of Gaussian and count responses (Scenario IV in Alexopoulos and Bottolo 
# (2021))

d <- Sim_GCRdata(m = 4, n = 100, p = 100, pFixed = 2, responseType = c("Gaussian", "binomial", 
                 "negative binomial", "negative binomial"), 
                 extras = list(rhoY = 0.8, rhoX = 0.7, fixedEffect = c(-0.5, 0.5), muEffect = 
                               c(1, rep(0.5, 3)), varEffect = c(1, rep(0.2, 3)), s1Lev = 0.05, 
                               s2Lev = 0.95, sdResponse = rep(1, 4), nTrial = 10, negBinomPar = 
                               c(0.5, 0.75)), 
                 seed = 28061971)

output <- BVS4GCR(d$Y, d$X, pFixed = 2, responseType = c("Gaussian", "binomial", 
                  "negative binomial", "negative binomial"), EVgamma = c(5, 3 ^2), 
                  niter = 250, burnin = 50, thin = 4, monitor = 50, seed = 280610971, 
                  nTrial = 10, negBinomParInit = rep(1, 2))


}
\references{
\insertAllCited{}
}
