% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Sim_GCRdata.R
\name{Sim_GCRdata}
\alias{Sim_GCRdata}
\title{Sim_GCRdata}
\usage{
Sim_GCRdata(
  m,
  n,
  p,
  pFixed,
  responseType,
  extras = list(rhoY = 0.8, rhoX = 0.7, fixedEffect = rep(1, pFixed), muEffect = rep(0,
    m), varEffect = rep(1, m), s1Lev = 0.1, s2Lev = 1, sdResponse = rep(1, m), nCat =
    NULL, cutPoint = NULL, nTrial = NULL, negBinomPar = NULL),
  std = TRUE,
  seed = 31122021
)
}
\arguments{
\item{m}{Integer, number of responses}

\item{n}{Integer, number of samples}

\item{p}{Integer, number of predictors}

\item{pFixed}{Integer, number of covariates (including the intercept). If \code{pFixed = 1} only the intercept will be simulated}

\item{responseType}{\code{m}-dimensional character vector specifying the response's type. Response's types currently supported are
\code{c("Gaussian", "binary", "ordinal", "binomial", "negative binomial")}}

\item{extras}{List specifying response-specific arguments. These are the one-lag correlation \code{rhoY} in the autoregressive
model used to simulate the responses, the correlation \code{rhoX} between the predictors (for both, see details in
\insertCite{Rothman2010;textual}{BVS4GCR}), \code{fixedEffect} the vector of the regression coefficients for the fixed
predictors (if simulated) where the first element is assumed to be the intercept, the mean \code{muEffect} and variance
\code{varEffect} of the Gaussian distribution used to simulate the regression coefficients, the probabilities \code{s1Lev} and
\code{s2Lev} utilized to induce sparsity on the regression coefficients, the standard deviation \code{sdResponse} of the response
variables, the number of categories \code{nCat} of each ordinal categorical response (if simulated), the cut-off points
\code{cutPoint} used to simulate the ordinal variables (if simulated), \code{nTrial} and \code{negBinomPar} the numer of trials in
the binomial experiment and the over-dispersion parameter of the negative binomial distribution. See also the examples below and
details in \insertCite{Alexopoulos2021;textual}{BVS4GCR}}

\item{std}{Logical parameter to standardise the predictors before the analysis. Default value is set at \code{TRUE}}

\item{seed}{Seed used to initialise the simulation}
}
\value{
The value returned is a list object \code{list(Y, X, B, R, par)}
\itemize{
\item{\code{Y}}{ \code{n} times \code{m} matrix of responses}
\item{\code{X}}{ \code{n} times (\code{pFixed} + \code{p}) matrix of covariates (including the intercept if simulated) and
predictors on which Bayesian Variable Selection is performed}
\item{\code{B}}{ \code{m} times (\code{pFixed} + \code{p}) matrix of regression coefficients}
\item{\code{R}}{ \code{m} times \code{m} correlation matrix used to simulate the responses}
\item{\code{extra}}{ List of all parameters used in the simulation \code{list(rhoY, rhoX, fixedEffect, muEffect, varEffect, 
       s1Lev, s2Lev, sdResponse, nCat, cutPoint, nTrial, negBinomPar, std, seed)} } }
}
\description{
Simulate data from the Gaussian Copula Regression model
}
\details{
The parameters \code{s1Lev} and \code{s2Lev} in the \code{extras} list can be specified following the fact that
((1 - \code{s2Lev}) * \code{p}) predictors are irrelevant for all \code{m} responses and that each relevant predictor is common
across (\code{s1Lev} * \code{m}) predictors \insertCite{Rothman2010}{BVS4GCR}. Type of responses currently supported: (continuous)
Gaussian, (discrete) binary, ordinal categorical and count (binomial and negative binomial distributions)
}
\examples{
# Example 1: Simulate a combination of Gaussian, binary and ordinal responses

d <- Sim_GCRdata(m = 4, n = 500, p = 30, pFixed = 1, responseType = c("Gaussian", "binary", 
                 "ordinal", "ordinal"), 
                 extras = list(rhoY = 0.8, rhoX = 0.7, fixedEffect = 1, muEffect = 
                               c(1, rep(0.5, 3)), varEffect = c(1, rep(0.2, 3)), s1Lev = 0.15, 
                               s2Lev = 0.95, sdResponse = rep(1, 4), nCat = c(3, 4), cutPoint = 
                               cbind(c(0.5, NA, NA), c(0.5, 1.2, 2))), 
                 seed = 28061971)

par(mfrow = c(2, 2))
hist(d$Y[,1], main = "Gaussian", xlab = expression(Y[1]), prob = TRUE, ylab = "")
barplot(prop.table(table(d$Y[, 2])), main = "binary", xlab = expression(Y[2]))
barplot(prop.table(table(d$Y[, 3])), main = "ordinal (3 categories)", xlab = expression(Y[3]))
barplot(prop.table(table(d$Y[, 4])), main = "ordinal (4 categories)", xlab = expression(Y[4]))

par(mfrow = c(2, 2))
plot(d$B[, 1], ylab=expression(beta[1]), main=expression(Y[1]), xlab="Pred.", pch=16, cex=.7)
plot(d$B[, 2], ylab=expression(beta[2]), main=expression(Y[2]), xlab="Pred.", pch=16, cex=.7)
plot(d$B[, 3], ylab=expression(beta[3]), main=expression(Y[3]), xlab="Pred.", pch=16, cex=.7)
plot(d$B[, 4], ylab=expression(beta[4]), main=expression(Y[4]), xlab="Pred.", pch=16, cex=.7)


# Example 2: As in Example 1 with more categories for the forth response 

d <- Sim_GCRdata(m = 4, n = 500, p = 30, pFixed = 1, responseType = c("Gaussian", "binary", 
                 "ordinal", "ordinal"), 
                 extras = list(rhoY = 0.8, rhoX = 0.7, fixedEffect = 1, muEffect = 
                               c(1, rep(0.5, 3)), varEffect = c(1, rep(0.2, 3)), s1Lev = 0.15, 
                               s2Lev = 0.95, sdResponse = rep(1, 4), nCat = c(3, 5), cutPoint = 
                               cbind(c(0.5, 1, NA, NA), c(0.5, 0.7, 1.5, 2))), 
                 seed = 28061971)

par(mfrow = c(2, 2))
hist(d$Y[, 1], main = "Gaussian", xlab = expression(Y[1]), prob = TRUE, ylab = "")
barplot(prop.table(table(d$Y[, 2])), main = "binary", xlab = expression(Y[2]))
barplot(prop.table(table(d$Y[, 3])), main = "ordinal (3 categories)", xlab = expression(Y[3]))
barplot(prop.table(table(d$Y[, 4])), main = "ordinal (5 categories)", xlab = expression(Y[4]))


# Example 3: Simulate a combination of Gaussian and count responses

d <- Sim_GCRdata(m = 4, n = 1000, p = 100, pFixed = 2, responseType = c("Gaussian", "binomial", 
                 "negative binomial", "negative binomial"), 
                 extras = list(rhoY = 0.8, rhoX = 0.7, fixedEffect = c(-0.5, 0.5), muEffect = 
                 c(1, rep(0.5, 3)), varEffect = c(1, rep(0.2, 3)), s1Lev = 0.05, s2Lev = 0.95, 
                 sdResponse = rep(1, 4), nTrial = 10, negBinomPar = c(0.5, 0.75)), 
                 seed = 28061971)

par(mfrow = c(2, 2))
hist(d$Y[,1], main = "Gaussian", xlab = expression(Y[1]), prob = TRUE, ylab = "")
hist(table(d$Y[, 2]), main = "binomial", xlab = expression(Y[2]))
hist(table(d$Y[, 3]), main = "negative binomial", xlab = expression(Y[3]))
hist(table(d$Y[, 4]), main = "negative binomial", xlab = expression(Y[4]))

}
\references{
\insertAllCited{}
}
